warning fix, and tiny fix of range around cursor for waterwell
git-svn-id: svn://svn.berlios.de/lincity-ng/trunk@1308 8226ee30-1aee-0310-a633-b68f58d1eb45

+++ src/lincity/lclib.cpp
@@ -229,7 +229,8 @@ void *lcalloc(size_t size)
     void *tmp;
     tmp = malloc(size);
     if (tmp == NULL) {
-        printf("couldn't malloc %d bytes!  Dying.\n", size);
+        /* prevent gcc warning on amd64: argument 2 has type 'long unsigned int' !!! */
+        printf("couldn't malloc %d bytes!  Dying.\n", (int) size);
         exit(-1);
     }
